BACKEND/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
python-pptx==0.6.23
requests==2.31.0
werkzeug==3.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
twilio==8.5.0
